chore: restructure src/test/ to mirror the source structure
- Create src/test/commands/ — ValidateNewProjectGAV.test.ts (was root)
- Move activation.test.ts → src/test/extension/ (was root)
- Rename VSCodeKaotoEditorChannelApi.test.ts →
  src/test/extension/VSCodeKaotoEditorChannelApiSettings.test.ts (was root)
- Rename src/test/helpers/ → src/test/utils/
- Move ClasspathRootFinder.test.ts → src/test/utils/ (was root)
- Create src/test/views/deployments/ and src/test/views/infrastructure/
  and move view tests into domain subfolders
- Update all import paths in moved files
