+++ src/test/extension/activation.test.ts
