+++ src/test/utils/DockerErrorDetector.test.ts
